Inital CSV file generation.

+++ config/config.go
@@ -7,70 +7,70 @@ type Config struct {
 	}
 
 	Time struct {
-		TicksPerDay uint
-		TotalDays   uint
+		TicksPerDay uint64
+		TotalDays   uint64
 	}
 
 	Arena ArenaConfig
 
 	Moral struct {
-		MajorMajorLow uint
-		MajorMinorLow uint
-		MinorMinorLow uint
+		MajorMajorLow uint64
+		MajorMinorLow uint64
+		MinorMinorLow uint64
 
-		RadiusMean   uint
-		RadiusStdDev uint
+		RadiusMean   uint64
+		RadiusStdDev uint64
 	}
 
 	Agent struct {
-		Civilian uint
-		Offender uint
-		Police   uint
+		Civilian uint64
+		Offender uint64
+		Police   uint64
 	}
 
 	Activity struct {
-		SleepMean   uint
-		SleepStdDev uint
+		SleepMean   uint64
+		SleepStdDev uint64
 
-		CountMean   uint
-		CountStdDev uint
+		CountMean   uint64
+		CountStdDev uint64
 	}
 
 	Workspace struct {
-		MajorMajorLow  uint
-		MajorMajorHigh uint
-		MajorMinorLow  uint
-		MajorMinorHigh uint
-		MinorMinorLow  uint
-		MinorMinorHigh uint
+		MajorMajorLow  uint64
+		MajorMajorHigh uint64
+		MajorMinorLow  uint64
+		MajorMinorHigh uint64
+		MinorMinorLow  uint64
+		MinorMinorHigh uint64
 	}
 
 	Economy struct {
-		PayPeriod uint
-		PayRate   uint
+		PayPeriod uint64
+		PayRate   uint64
 
-		Unemployment uint
-		HiringRate   uint
-		FiringRate   uint
+		Unemployment uint64
+		HiringRate   uint64
+		FiringRate   uint64
 
-		WealthMean   uint
-		WealthStdDev uint
+		WealthMean   uint64
+		WealthStdDev uint64
 	}
 
 	Offender struct {
-		Amount   uint
-		Cooldown uint
+		Amount   uint64
+		Cooldown uint64
 	}
 }
 
 // ArenaConfig represents the configuration of the arena.
 type ArenaConfig struct {
-	Width  uint
-	Height uint
+	Width  uint64
+	Height uint64
 
-	MajorX uint
-	MajorY uint
+	MajorX uint64
+	MajorY uint64
 
-	MinorWeight uint
-	MajorWeight uint
+	MinorWeight uint64
+	MajorWeight uint64
 }
